Externalize paragraph paths calculation to business file
